fixed circ dependencies and moved bundle tx to block class
