+++ src/block/__init__.py
@@ -1,1 +1,1 @@
-from src.block.block import Block
+from src.block.block import Block, bundle_tnx
