+++ src/core/api.py
@@ -4,9 +4,9 @@ import os
 
 from Crypto.Hash import SHA256
 from src.persist import save_block, save_unverified_transaction, save_verified_transaction, read_block, save_utxo
-from src.block.block import Block
+from src.block.block import Block, genesis_block, bundle_tnx
 from src.configuration import Configuration
-from src.transaction import Transaction, bundle_tnx
+from src.transaction import Transaction
 from src.wallet import get_public_key, get_private_key
 
 
@@ -45,6 +45,7 @@ def create_transaction(recipient, amount):
         tx.unlock_inputs(get_private_key(), get_public_key("string"))
         save_verified_transaction(tx.get_transaction_id(), tx.get_data())
     except ValueError as e:
+        print(e)
         raise ValueError("INSUFFICIENT FUNDS")
 
 def get_block(block_hash):
@@ -58,24 +59,40 @@ def get_block(block_hash):
 def init_configuration():
     """
     This will instantiate the config class
+    and check if genesis block is created
     :return: configuration object
     """
+    conf = Configuration()
 
-    # TODO: Genesis block should be created here
-    # TODO: Then call finc_incoming_utxos
-    return Configuration()
+    block_height = conf.get_conf("height")
+    if block_height == 0:
+        b = genesis_block()
+        cwd = os.getcwd()
+        try:
+            os.makedirs(os.path.join(cwd, r'data'))
+        except OSError as e:
+            pass
+        save_block(b)
+        conf.increment_height()
+        find_incoming_utxos(b.block_hash(), b.transactions, True)
 
-def find_incoming_utxos(block_hash, transactions):
+    return conf
+
+def find_incoming_utxos(block_hash, transactions, isGenesis=False):
     """
     Iterates through all the outputs and looks for any directed to user's wallet.
     If found, save to the utxo pool
     :return:
     """
     myAddress = SHA256.new(get_public_key("string").encode()).hexdigest()
+    for tnx_id, tnx_info in transactions.items():
+        # deserialize transaction
+        tnx_payload = tnx_info
+        tnx_payload["transaction_id"] = tnx_id
+        tnx = Transaction(payload=tnx_payload)
 
-    for tnx in transactions:
         for index in range(len(tnx.outputs)):
-            if tnx.outputs[index]["address"] == myAddress and len(tnx.inputs) > 0:
+            if tnx.outputs[index]["address"] == myAddress and not isGenesis:
                 save_utxo(tnx.get_transaction_id(), index, block_hash, tnx.outputs[index]["amount"])
-            elif tnx.outputs[index]["address"] == myAddress and len(tnx.inputs) == 0:
+            elif tnx.outputs[index]["address"] == myAddress and isGenesis:
                 save_utxo(tnx.get_transaction_id(), -1, block_hash, tnx.outputs[index]["amount"])
